% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian_code_final.R
\name{bayesian_modules}
\alias{bayesian_modules}
\title{bayesian modules function}
\usage{
bayesian_modules(R, select = 1e+05)
}
\arguments{
\item{R}{is a list of correlation matrices of length L (the number of type 1 studies). The correlation matrices are of size pxp and are all of the same size.}

\item{select}{is the number of items we want to select in order to estimate the parameters, default value is set to 10^5}
}
\value{
param a list of the estimated parameters

med_theta the posterior median of theta

mean_theta the posterior mean of theta

sd_theta the posterior standard deviation of theta

post_theta the posterior probability that theta is 0

gene_names the names of the genes included in the analysis
}
\description{
This function computes the common correlation matrix from a set of correlation matrices
}
