% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian_code_final.R
\name{consensus_cluster}
\alias{consensus_cluster}
\title{consensus_cluster function}
\usage{
consensus_cluster(R, kmin = 2, kmax, clusterAlg = "hclust_ward", B = 10,
  prop_resamp = 0.8)
}
\arguments{
\item{R}{a correlation matrix}

\item{kmin}{the minimum number of clusters}

\item{kmax}{the maximum number of clusters}

\item{clusterAlg}{the clustering algorithm, one of the following kmeans, hclust_ward,hclust_average,hclust_complet,hclust_single}

\item{B}{the number of resampling steps}

\item{prop_resamp}{the proportion of the data that we want to resample}
}
\value{
best k (number of clusters) chosen by this method.
}
\description{
This function performs consensus clustering.
}
