% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian_code_final.R
\name{EB_module}
\alias{EB_module}
\title{EB_module function}
\usage{
EB_module(R, gap = FALSE, k = 500, scattered = FALSE,
  method_scattered = c("GPD", "kmeans"), Kmin = NULL, Kmax = NULL,
  Nperm = 100, dicho_f = 2)
}
\arguments{
\item{R}{is a list of lenght L1 (number of studies) of correlation matrices of the same size}

\item{gap}{whether to use the gap statistic to determine the optimal number of clusters}

\item{k}{if gap is TRUE then k=NULL otherwise, k is an integer indicating the number of desired modules (scattered genes do not count as a module)}

\item{scattered:}{bolean whether the function should look for scattered genes before looking for modules.}
}
\value{
common_r the common correlation matrix estimated through the empirical bayesian model

modules a list of modules with the scattered genes grouped inside a module (called -1)

bestk the optimal number of modules according to a modified version of the GAP statistic (if gap is set to TRUE)
}
\description{
This function is the main function of the package. It computes the common gene modules for a collection of correlation matrices
}
