% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian_code_final.R
\name{fisher}
\alias{fisher}
\title{fisher function}
\usage{
fisher(r)
}
\arguments{
\item{r}{is a vector of correlations}
}
\value{
a vector of normaly distributed data
}
\description{
The fisher transformation of a vector of correlations
}
