% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian_code_final.R
\name{fisher_inv}
\alias{fisher_inv}
\title{fisher_inv function}
\usage{
fisher_inv(z)
}
\arguments{
\item{z}{a normal vector}
}
\value{
a vector of correlations.
}
\description{
Inverse of the fisher transformation,
}
