% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian_code_final.R
\name{GAP_dicho}
\alias{GAP_dicho}
\title{GAP_dicho function}
\usage{
GAP_dicho(R, Kmin, Kmax, Nperm = 100, method = "ward.D",
  bestk_method = "Tib", dicho_f = 2)
}
\arguments{
\item{R}{is a correlation matrix}

\item{Kmin}{gives the minimum number of clusters that we want to obtain}

\item{Kmax}{gives the maximum number of clusters that we want to obtain}

\item{Nperm}{is the number of resampling that we want to do}

\item{method}{can only be one of the proposed clustering algorithm. it must return a list with one of the elements named "cluster", which is a vector of length p (number of genes) and containing the id of the cluster for each gene}

\item{bestk_method}{the way to compute the best k as describe in the GAP package}

\item{dicho_f}{the factor for the dichotomy, A small value will lead to several rounds.}
}
\value{
bestk the optimal number of clusters.
}
\description{
This function is a modified version of the GAP statistic for very large datasets.
}
