% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian_code_final.R
\name{hclust_fun}
\alias{hclust_fun}
\title{hclust_fun function}
\usage{
hclust_fun(R, k, met = c("ward", "average", "complete", "single"))
}
\arguments{
\item{R}{a correlation matrix}

\item{k}{the number of clusters}

\item{met}{the method for hierarchical clustering, one of ward, average, complete or single}
}
\value{
a vector where each gene is attributed a cluster
}
\description{
function that performs hierarchical clustering and outputs the right format for our gap function
}
