% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian_code_final.R
\name{kmeans_fun}
\alias{kmeans_fun}
\title{kmeans_fun function}
\usage{
kmeans_fun(R, k, met = NULL)
}
\arguments{
\item{R}{a correlation matrix}

\item{k}{the number of clusters}

\item{met}{default parameter set to NULL}
}
\value{
a vector where each gene is attributed a cluster
}
\description{
function that performs kmeans clustering and outputs the right format for our gap function
}
