% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian_code_final.R
\name{lik_mixgpd_onexi}
\alias{lik_mixgpd_onexi}
\title{lik_mixgpd_onexi function}
\usage{
lik_mixgpd_onexi(param, x)
}
\arguments{
\item{sigma1, }{parameter of the first GPD distribution}

\item{sigma2, }{parameter of the second GPD distribution}

\item{xi}{common shape parameter for both distributions}

\item{p}{proportion of mixture of each distribution}
}
\value{
value of the likelihood
}
\description{
This function computes the likelihood  of a mixture of two GPD distributions with a common xi parameter
}
