% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian_code_final.R
\name{likelihood_logit}
\alias{likelihood_logit}
\title{likelihood_logit function}
\usage{
likelihood_logit(param, Z)
}
\arguments{
\item{param}{is a vector of length L+2, where the L first vectors are sigma (the standard deviations for each of the different type1 studies), 
q=logit(p)=param[L+1], where p is the probability that theta is 0 (trans(q)=p) and 
tau=param[L+2], the sd of theta.}

\item{Z}{a vector of z values (see bayesian_modules)}
}
\value{
the value of the likelihood
}
\description{
This function computes the likelihood for the empirical bayesian model. It uses the logit transformation to estimate p (simpler)
}
