% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian_code_final.R
\name{logit}
\alias{logit}
\title{logit function}
\usage{
logit(p)
}
\arguments{
\item{p}{a vector with values between 0 and 1 which we want to take logit.}
}
\description{
This function computes the logit of a quantity
}
\examples{
logit(runif(100,0,1))
}
\keyword{logit}
