% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian_code_final.R
\name{maxmean}
\alias{maxmean}
\title{maxmean function}
\usage{
maxmean(Z, R)
}
\arguments{
\item{Z}{are the Z to summarize, so corresponding to one module}

\item{R}{is the correlation matrix corresponding to Z.}
}
\value{
the summarized version of z for the module
}
\description{
function to summarize the information of a Z vector into a module
}
