% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian_code_final.R
\name{median_theta_function}
\alias{median_theta_function}
\title{median_theta_function}
\usage{
median_theta_function(mean, sd, p_z)
}
\arguments{
\item{mean}{the mean of theta}

\item{sd}{the standard deviation of theta}

\item{p_z}{the posterior probability of being 0}
}
\value{
the posterior median of theta
}
\description{
Function to compute the posterior median of theta
}
