% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian_code_final.R
\name{RandIndex}
\alias{RandIndex}
\title{RandIndex function}
\usage{
RandIndex(clusX, clusY)
}
\arguments{
\item{clusX}{and clusY lists of length "number of clusters" and trying to cluster the same number of genes}
}
\value{
the value of the rand index
}
\description{
This function computes the rand index for two clusters clusX and clusY.
}
