% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian_code_final.R
\name{rank_mean}
\alias{rank_mean}
\title{rank_mean function}
\usage{
rank_mean(R)
}
\arguments{
\item{R}{the ranks to be summarized into one value}
}
\description{
function to summarize ranks into modules
}
