% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian_code_final.R
\name{scattered_by_kmeans}
\alias{scattered_by_kmeans}
\title{scattered_by_kmeans function}
\usage{
scattered_by_kmeans(T)
}
\arguments{
\item{T}{the statistic from which we want to discriminate between scattered and non scattered genes}
}
\value{
scattered the names of the scattered genes
}
\description{
Function to find scattered genes by kmeans with k=2
}
