% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian_code_final.R
\name{scattered_extremes_onexi}
\alias{scattered_extremes_onexi}
\title{scattered_extremes_onexi function}
\usage{
scattered_extremes_onexi(T, quantile_T = 0.9, hessian = FALSE)
}
\arguments{
\item{T}{statistic whose tail/most extremes values should be the scattered genes for the most part}

\item{quantile_T}{is the threshold that we choose for the tail}

\item{hessian}{bolean to indicate whether we want to compute the standard deviation of the parameters}
}
\value{
scattered gives the name of the scattered genes, the value of the parameters and if hessian is TRUE, their standard error
}
\description{
This function is used to discriminate between scattered and non scattered genes. It fits a mixture of GPD distributions with a common xi.
}
\details{
here we consider that xi1=xi2 for a better estimation of this parameter
}
