% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian_code_final.R
\name{trans}
\alias{trans}
\title{trans function}
\usage{
trans(q)
}
\arguments{
\item{q}{parameter that we want to apply inverse logit on}
}
\description{
this function is the inverse of the logit. Useful for estimating p in the likelihood optimization
}
\examples{
trans(rnorm(100,0,1))
}
\keyword{logit}
