% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian_code_final.R
\name{Wk}
\alias{Wk}
\title{Wk function}
\usage{
Wk(R, cluster)
}
\arguments{
\item{R}{is a correlation matrix}

\item{cluster}{is the results of a clustering procedure, a vector of length number of genes indicating the cluster to which the corresponding gene belongs.}
}
\value{
Wk the average sum of pairwise distances across clusters
}
\description{
function to compute the average sum of pairwise distances across clusters as defined in the paper from Tibshirani
}
